index.html
This HTML file represents a webpage for ARTSYSNAPS86, a photography studio. The page features a sleek black background with white text and is centered for a modern and stylish look. It introduces the studio, emphasizing its dedication to blending outdoor and studio photography seamlessly. Visitors are encouraged to explore the photo gallery, order with free delivery, and learn about competitive pricing. The founder, Punniyakotti S, is introduced as an aspiring undergraduate with a deep passion for photography. Contact information is provided, making it easy for potential clients to reach out. This webpage effectively presents the studio's services and showcases its creative approach to photography.